Move controller and view classes to sanapuuro.ui package
