+++ Sanapuuro/src/main/java/sanapuuro/AiPlayer.java
@@ -1,5 +1,6 @@
 package sanapuuro;
 
+import sanapuuro.ui.Controller;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
