+++ Sanapuuro/src/main/java/sanapuuro/ui/GameView.java
@@ -3,10 +3,14 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package sanapuuro;
+package sanapuuro.ui;
 
+import sanapuuro.ui.View;
+import sanapuuro.ui.Controller;
+import sanapuuro.ui.ControllerListener;
 import java.util.Collections;
 import java.util.List;
+import sanapuuro.Player;
 import sanapuuro.grid.Grid;
 import sanapuuro.letters.Letter;
 import sanapuuro.letters.Letters;
